chore(api-service): add e2e test to cover the digest eventCount, and events variables with filters used
